\name{sampleGeneNames}
\alias{sampleGeneNames}
\docType{data}
\title{
Sample gene names
}
\description{
A list of gene names used to generate permutation of names in \code{computeRandomizedResult} method.
}
\usage{data(sampleGeneNames)}
\format{
 A factor of gene names
}
\examples{
data(sampleGeneNames)
}
\keyword{datasets}
