\name{samplePatientExpressionMatrix}
\alias{samplePatientExpressionMatrix}
\docType{data}
\title{
Sample patient expression matrix
}
\description{
A real-value matrix representing gene expressions in patients.
}
\usage{data(samplePatientExpressionMatrix)}
\format{
  A two-dimensional real value matrix. The row names are patients, and collumn names are genes. The entry samplePatientExpressionMatrix[i,j] is the gene expression values of gene j in patient i.
}
\examples{
data(samplePatientExpressionMatrix)
}
\keyword{datasets}
