\name{sampleRandomDriversResult}
\alias{sampleRandomDriversResult}
\docType{data}
\title{
Sample Result from computeRandomizedResult
}
\description{
A sample result from running \code{computeRandomizedResult} with numberOfRandomTests = 20.
}
\usage{data(sampleRandomDriversResult)}
\format{
  A list of predicted drivers from different runs.
}
\seealso{
\code{computeRandomizedResult}
}
\examples{
data(sampleRandomDriversResult)
}
\keyword{datasets}
